Factor out ProfileForms
